Move some data to StaticData class
